requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
gTTS==2.3.1
idna==3.4
praw==7.7.0
prawcore==2.3.0
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
